crop controller to seperate file
