+++ lib/src/crop.dart
@@ -8,6 +8,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 import 'package:vector_math/vector_math_64.dart' as vm;
 
+part 'crop_controller.dart';
+
 /// Used for cropping the [child] widget.
 class Crop extends StatefulWidget {
   final Widget child;
@@ -344,103 +346,6 @@ class _CropState extends State<Crop> with TickerProviderStateMixin {
     _controller.dispose();
     widget.controller.removeListener(_reCenterImage);
     super.dispose();
-  }
-}
-
-typedef _CropCallback = Future<ui.Image> Function(double pixelRatio);
-
-/// The controller used to control the rotation, scale and actual cropping.
-class CropController extends ChangeNotifier {
-  double _aspectRatio = 1;
-  double _rotation = 0;
-  double _scale = 1;
-  Offset _offset = Offset.zero;
-  _CropCallback? _cropCallback;
-
-  /// Gets the current aspect ratio.
-  double get aspectRatio => _aspectRatio;
-
-  /// Sets the desired aspect ratio.
-  set aspectRatio(double value) {
-    _aspectRatio = value;
-    notifyListeners();
-  }
-
-  /// Gets the current scale.
-  double get scale => max(_scale, 1);
-
-  /// Sets the desired scale.
-  set scale(double value) {
-    _scale = max(value, 1);
-    notifyListeners();
-  }
-
-  /// Gets the current rotation.
-  double get rotation => _rotation;
-
-  /// Sets the desired rotation.
-  set rotation(double value) {
-    _rotation = value;
-    notifyListeners();
-  }
-
-  /// Gets the current offset.
-  Offset get offset => _offset;
-
-  /// Sets the desired offset.
-  set offset(Offset value) {
-    _offset = value;
-    notifyListeners();
-  }
-
-  /// Gets the transformation matrix.
-  Matrix4 get transform => Matrix4.identity()
-    ..translate(_offset.dx, _offset.dy, 0)
-    ..rotateZ(_rotation)
-    ..scale(_scale, _scale, 1);
-
-  /// Constructor
-  CropController({
-    double aspectRatio = 1.0,
-    double scale = 1.0,
-    double rotation = 0,
-  }) {
-    _aspectRatio = aspectRatio;
-    _scale = scale;
-    _rotation = rotation;
-  }
-
-  double _getMinScale() {
-    final r = vm.radians(_rotation % 360);
-    final rabs = r.abs();
-
-    final sinr = sin(rabs).abs();
-    final cosr = cos(rabs).abs();
-
-    final x = cosr * _aspectRatio + sinr;
-    final y = sinr * _aspectRatio + cosr;
-
-    final m = max(x / _aspectRatio, y);
-
-    return m;
-  }
-
-  /// Capture an image of the current state of this widget and its children.
-  ///
-  /// The returned [ui.Image] has uncompressed raw RGBA bytes, will have
-  /// dimensions equal to the size of the [child] widget multiplied by [pixelRatio].
-  ///
-  /// The [pixelRatio] describes the scale between the logical pixels and the
-  /// size of the output image. It is independent of the
-  /// [window.devicePixelRatio] for the device, so specifying 1.0 (the default)
-  /// will give you a 1:1 mapping between logical pixels and the output pixels
-  /// in the image.
-  Future<ui.Image> crop({double pixelRatio = 1}) {
-    if (_cropCallback == null) {
-      throw Exception('Crop callback is not set');
-    }
-
-    return _cropCallback!.call(pixelRatio);
   }
 }
